Updated Hex ID in DataStructure to be an array of bytes instead of a string.
Properties which lie in a data structure's introducer field now implement a method which automatically updates the raw introducer data.

Added an "Encode" (save) method in AFPFile.cs to write the fields back to a byte stream, including the 0x5A prefix.

Wrote unit tests to ensure that the decoding logic still works as expected, along with testing the new encoding method, and comparing the result to the original raw file.